Reuse StoriesContainer
