+++ src/containers/StoriesContainer.jsx
@@ -1,17 +1,8 @@
-import { useEffect, useState } from 'react'
-import { getStoryIds } from '../services/hackerNewsApi'
 import { Story } from '../components/Story'
 import ShowMoreButton from '../components/ShowMoreButton'
 import { STORY_PAGE_SIZE } from '../constants'
 
-const StoriesContainer = () => {
-  const [storyIds, setStoryIds] = useState([])
-  const [pageNumber, setPageNumber] = useState(1)
-
-  useEffect(() => {
-    getStoryIds().then((res) => setStoryIds(res))
-  }, [pageNumber])
-
+const StoriesContainer = ({ storyIds, pageNumber, setPageNumber }) => {
   return (
     <div className='StoriesContainer'>
       {storyIds.slice(0, pageNumber * STORY_PAGE_SIZE).map((id, index) => (
